refactor: clean up and separate out providers
