+++ composer.json
@@ -3,15 +3,13 @@
     "description": "ADDON_DESCRIPTION",
     "type": "wordpress-plugin",
     "license": "GPL-2.0-or-later",
-    "require": {},
     "require-dev": {
-        "phpunit/phpunit": "^5",
-        "yoast/phpunit-polyfills": "^1.1.0",
         "squizlabs/php_codesniffer": "^3.5",
         "phpcompatibility/phpcompatibility-wp": "^2.1",
         "wp-coding-standards/wpcs": "^2.3",
         "phpcompatibility/php-compatibility": "^9.3",
-        "dealerdirect/phpcodesniffer-composer-installer": "^0.7.0"
+        "dealerdirect/phpcodesniffer-composer-installer": "^0.7.0",
+        "yoast/phpunit-polyfills": "^4.0"
     },
     "scripts": {
         "test": "./vendor/bin/phpunit --colors --stop-on-failure",
@@ -36,7 +34,7 @@
     },
     "config": {
         "platform": {
-            "php": "7.2"
+            "php": "7.4.0"
         },
         "allow-plugins": {
             "dealerdirect/phpcodesniffer-composer-installer": true
