--- src/Domain/resources/css/frontend/frontend.scss
@@ -1,1 +1,0 @@
-// definitions
