--- src/Domain/resources/js/admin/admin.ts
@@ -1,3 +1,0 @@
-import {imageUrl} from '../utils/helpers';
-
-console.log('[ADDON_ID] admin script - imageUrl: ', imageUrl('give-logo.png'));
