--- src/Domain/resources/js/frontend/frontend.js
@@ -1,3 +1,0 @@
-import {imageUrl} from '../utils/helpers';
-
-console.log('[ADDON_ID] frontend script - imageUrl: ', imageUrl('give-logo.png'));
