--- src/Domain/resources/js/utils/helpers.ts
@@ -1,8 +1,0 @@
-import {IGiveAddon} from './interfaces';
-
-declare global {
-    interface Window {
-        GiveAddon: IGiveAddon;
-    }
-}
-export const imageUrl = (filename) => `${window.GiveAddon.imageUrl}${filename}`;
