+++ src/Settings/resources/js/components/Form/Form.module.scss
