+++ src/Settings/resources/js/components/Icons.module.scss
