+++ src/Settings/resources/js/components/Icons.tsx
