+++ src/Settings/resources/js/components/Modal.module.scss
