+++ src/Settings/resources/js/index.tsx
@@ -1,13 +1,7 @@
 import React from 'react';
 import {createRoot} from 'react-dom/client';
 import App from './App';
-import {IGiveAddon} from '../../utils/interfaces';
 
-declare global {
-    interface Window {
-        GiveAddon: IGiveAddon;
-    }
-}
 
 const container = document.getElementById('ADDON_ID-settings-page-app');
 const root = createRoot(container!);
