+++ src/Settings/resources/js/utils/getWindowData.ts
@@ -1,0 +1,7 @@
+import {IGiveAddon} from './interfaces';
+
+declare const window: {
+    GiveAddon: IGiveAddon;
+} & Window;
+
+export const getWindowData = () => window.GiveAddon;
