+++ src/Settings/resources/js/utils/helpers.ts
@@ -1,0 +1,6 @@
+import {IGiveAddon} from './interfaces';
+
+import {getWindowData} from './getWindowData';
+
+
+export const imageUrl = (filename) => `${getWindowData().GiveAddon.imageUrl}${filename}`;
