+++ webpack.config.js
@@ -17,11 +17,7 @@ module.exports = {
         GiveAddonFormBuilderExtension: srcPath('FormExtension/FormBuilder/resources/js/index.tsx'),
         GiveAddonFormBuilderExtensionGlobalStyle: srcPath('FormExtension/FormBuilder/resources/css/index.scss'),
         GiveAddonOffSiteGateway: srcPath('OffSiteGateway/Gateway/resources/OffSiteGateway.tsx'),
-        settings: srcPath('Domain/resources/js/admin/settings-page-app/index.tsx'),
-        adminJs: srcPath('Domain/resources/js/admin/index.tsx'),
-        adminCss: srcPath('Domain/resources/css/admin/index.scss'),
-        frontendJs: srcPath('Domain/resources/js/frontend/index.tsx'),
-        frontendCss: srcPath('Domain/resources/css/frontend/index.scss'),
+        settings: srcPath('Settings/resources/js/admin/index.tsx'),
     },
 };
